+++ lib/src/constants/keyvalue.keys.dart
@@ -1,0 +1,4 @@
+class KeyValueKeys {
+  static const String lastPeriodicTaskExecution =
+      'last_periodic_task_execution';
+}