+++ lib/src/database/daos/mediafiles.dao.dart
@@ -121,6 +121,7 @@ class MediaFilesDao extends DatabaseAccessor<TwonlyDB>
           ..where(
             (t) => (t.uploadState.equals(UploadState.initialized.name) |
                 t.uploadState.equals(UploadState.uploadLimitReached.name) |
+                t.uploadState.equals(UploadState.uploading.name) |
                 t.uploadState.equals(UploadState.preprocessing.name)),
           ))
         .get();