+++ lib/src/database/daos/receipts.dao.dart
@@ -4,6 +4,7 @@ import 'package:hashlib/random.dart';
 import 'package:twonly/src/database/tables/messages.table.dart';
 import 'package:twonly/src/database/tables/receipts.table.dart';
 import 'package:twonly/src/database/twonly.db.dart';
+import 'package:twonly/src/services/api/mediafiles/upload.service.dart';
 import 'package:twonly/src/utils/log.dart';
 
 part 'receipts.dao.g.dart';
@@ -33,6 +34,7 @@ class ReceiptsDao extends DatabaseAccessor<TwonlyDB> with _$ReceiptsDaoMixin {
           type: const Value(MessageActionType.ackByUserAt),
         ),
       );
+      await handleMediaRelatedResponseFromReceiver(receipt.messageId!);
     }
 
     await (delete(receipts)