+++ lib/src/services/api/server_messages.dart
@@ -63,6 +63,7 @@ Future<void> handleServerMessage(server.ServerToClient msg) async {
       ..response = response;
 
     await apiService.sendResponse(ClientToServer()..v0 = v0);
+    globalGotMessageFromServer = true;
   });
 }