+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.98
 
 - New: Groups can now collect flames as well
+- New: Background execution to pre-load messages 
 - New: Adds a link if the image contains a QR code
 - Improve: Video compression with progress updates
 - Improve: Show message "Flames restored"